# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Go WhatsApp API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /instances:
    get:
      tags:
        - Instances
      summary: Listar instâncias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instance'
        '401': { description: Não autorizado }
        '403': { description: Token inválido }
    post:
      tags:
        - Instances
      summary: Criar instância
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstance'
      responses:
        '201': { description: Criada }
        '401': { description: Não autorizado }
        '403': { description: Token inválido }
  /instances/{name}:
    delete:
      tags:
        - Instances
      summary: Deletar instância
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletada
          content:
            application/json:
              schema:
                type: object
                properties:
                  Details:
                    type: string
        '401': { description: Não autorizado }
        '403': { description: Token inválido }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Token de autenticação Bearer. Pode ser:
        - Token específico da instância (obtido ao criar a instância)
        - Token master global (configurado via API_MASTER_TOKEN no .env)
  schemas:
    Instance:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        webhookUrl: { type: string, format: uri }
        token: { type: string }
        status: { type: string, enum: [pending_qr, connected, disconnected, logged_out, active] }
        webhook:
          $ref: '#/components/schemas/InstanceWebhookConfig'
        settings:
          $ref: '#/components/schemas/InstanceSettings'
    InstanceSettings:
      type: object
      properties:
        rejectCall: { type: boolean }
        msgCall: { type: string }
        groupsIgnore: { type: boolean }
        alwaysOnline: { type: boolean }
        readMessages: { type: boolean }
        readStatus: { type: boolean }
        syncFullHistory: { type: boolean }
    InstanceWebhookConfig:
      type: object
      properties:
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string }
        headers:
          type: object
          additionalProperties: { type: string }
        enabled: { type: boolean }
        byEvents: { type: boolean }
        base64: { type: boolean }
...
    CreateInstance:
      type: object
      required: [instanceName, webhookUrl, token]
      properties:
        instanceName: { type: string }
        webhookUrl: { type: string, format: uri }
        token: { type: string }
...